coding

+++ src/main/java/com/example/mytask/repository/UserRepository.java
@@ -1,12 +1,11 @@
 package com.example.mytask.repository;
 
 import com.example.mytask.model.User;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-  User findById(int id);
+  User findById(long id);
 }
